renamed all Page1 Articles with PascalCase
